+++ ref_man/__main__.py
@@ -10,7 +10,7 @@ def main():
     parser = argparse.ArgumentParser("ref-man")
     parser.add_argument("--no-threaded", dest="threaded", action="store_false",
                         help="Whether flask server should be threaded or not")
-    parser.add_argument("--host", "-h", default="localhost",
+    parser.add_argument("--host", default="localhost",
                         help="host on which to bind the python server")
     parser.add_argument("--port", "-p", type=int, default=9999,
                         help="Port to bind to the python server")
@@ -53,7 +53,7 @@ def main():
               "proxy_everything_port": args.proxy_everything_port,
               "data_dir": args.data_dir,
               "local_pdfs_dir": args.local_pdfs_dir,
-              "remote_pdfs_dir": args.local_pdfs_dir,
+              "remote_pdfs_dir": args.remote_pdfs_dir,
               "remote_links_cache": args.remote_links_cache,
               "batch_size": args.batch_size,
               "chrome_debugger_path": args.chrome_debugger_path,